[FIX] The block() and unblock() methods in the EditorManager.stopManagerListener did not prevent the sending of TextEditActivities
The method textAboutToBeChanged() in the EditorManager class gets called
right before TextEdits happen. This methods toggles the creation of
TextEditActivities. It won't send any TextEditActivities if the
hasWriteAccess property is set to false.

The only way this attribute is changed during a running session is by a
call of permissionChanged(), but it should also be changed when the user
is blocked by the StopManager.

This patch prevents the sending of TextEditActivites while the session
is stopped via the StopManager.

Change-Id: I2cd3515d5dada11a506fd660cbd371acd639eeba
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/441
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>