[UI] Only SessionHost can invite other users
This is necessary to prevent inconsistencies, which can still occur when
a non-host invites another user. If any JupiterActivities get lost
during an invitation, the JupiterServer needs to be reset for the
invited users. This can only be done by the host.

A consecutive patch could implement this function so that the non-host
can recommend a user to invite to the host.

If the invitation process is stable enough so that a non-host could
invite without causing inconsistencies, this feature could be activated 
again.

Change-Id: Ieda033a627fcf0c2a137ecccdf0a362b68ad767f

Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/380
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>