[INTERNAL|UI] do not use progress monitors in the network layer
Removed the usage of progress monitors for various util
methods from the RosterUtils class.

Changed monitor status from deterministic to unknown as
this process cannot be measured anyway.

Removed the cancellation because either the process
will be somewhat fast so the user cannot cancel it
or it will normally hang in the some parts
of the smack logic where cancellation is not even
possible.


Change-Id: Idff0b5d390f1be328965265cd6e35a19ed6ffdb2
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/416
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>