[INTERNAL] removed message queuing
Either the feature was never fully implemented or
was broken by a patch. Currently messages for
a participant that is not available are just stored
but never send again.

As this feature does not make sense anyway because when
the connection is lost, the DataTransferManager will automatically close
all existing connections, it is the best option to remove the left over
artifacts of this feature.

Change-Id: I405044e715a915c1c29a60aacf027164829438c9
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/477
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>