[INTERNAL] do not proceed the invitation if the remote side has no Saros support
The querying for Saros support is SOMEWHAT important because: it 
returns a resource qualified JID and thats the JID we use for the
whole session. Formerly if this process fails and the user decided to
continue a resource qualified JID was build.

This self build JID will not match the JID the user is logged in
on several XMPP severs.

E.G on GTALK XMPP servers your JID will be like
xxx@googlemail.com/SarosAB322 or something like that.

To keep it simple, just abort the invitation process if the remote
contact "may" not support Saros.

Change-Id: I73017a71e1dff8f0a68f0159c630527260a94353
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/516
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>