[INTERNAL] refactored SarosUI usage
The SarosUI component was injected into the network layer
and then passed as argument to the upper layers.

The reason for this behavior was that this component
is needed by the SarosSessionManager but the SarosUI
also need the SarosSessionManager and so the PICO
Container could not resolve the dependencies.

Removed all needed components from the SarosUI
as they can be retrieved in the wizards with the SarosPluginContext
method.

Change-Id: Icc433e25bdc3abe395aac30aa11393078466d5e9
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/564
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>