[GUI] inform the user that its selected color is already taken
It is possible that another user had already chosen the same
color when the user finished the wizard for changing its color.

Give it at least a try to inform the user that this color
is no longer available (it is still possible that a color 
change will not be recognized as the host
does not inform us if the color is already in use).

Change-Id: I84908a152a1343dc7d83318914d3f096f997619c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/567
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>