[STF] let the framework handle thread termination
After investigating the test code base it seems that
most developers tend to copy and paste code or reuse
an existing test case and fit it to their needs.

Working with threads is dangerous as it can crash
a whole regression. Added methods to the StfTestCase
base class the handle the correct termination of the
created threads.

If a thread could not be terminated after a test class all 
further test cases will automatically fail (depends on the used
class loader). 

Change-Id: Ia822286d99eab4c7d9277aabc999e18d1d3420a1
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/571
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>