[GUI] changed the progress report for an outgoing project negotiation
The original progress report was a little bit sloppy.

The progress bar was divided mainly in 50% hard disk stuff
and 50% network stuff.

Depending on the hardware of the user one of those two section
can be outstanding fast or amazingly slow.

Changed the progress report so that the progress bar now resets
after every long running operation to indicate that a new major
task has been started in the current project negotiation.


Change-Id: I052717ef21e8bcd3b24430ea48f244b282c8ce74
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/576
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>