[INTERNAL|GUI] disabled copy feature
After investigating the code, it seems that either
the partial sharing feature or multiple project feature
had broken this functionality. The IncomingProjectNegotiation
has still the code to make a copy of a project but lacks
of the required arguments.

As this feature currently does not work disable it by
hiding the necessary GUI fields.

Change-Id: Ie2a6cf1e2b8397e78746c416df4055e660436a21
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/577
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>