[STF] fixed a Superbot method
Changes from Alexander Waldmann introduced a new feature
that automatically displays a dialog if the host is currently
alone in a session. After this feature I changed the STFController
to auto accept the dialog which will then close the current session.

As per design all SuperBot methods that will return a boolean
will not throw any exception, so this bug is normally not noticed
as the method still returns the correct result.

The only time were you are able to notice this bug is in the extra 
amount of time (currently 5 seconds) that is spend in the affected
method before it returns.

Change-Id: Ia508036d5c272b680f6b1b34ff3c0db9ac91cc26
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/578
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>