[NOP] always send the correct session ID
Send the session ID the project negotiation process was created
with.

Note: This currently does not fix anything.

If a session is closed the session termination packet will mostly
send first before the cancellation message is send. This
may result in a state where this cancellation message is silently
dropped because it is a session packet and the remote side is not in
a session anymore.

Change-Id: I199c3e3c7483c6e033cca2757ac61190afdf4bfb
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/584
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>