[INTERNAL|GUI|STF] always show the exit session dialog if left alone in session
The STF highly depends on the automatically close of a session that
is executed by this handler. Stack traces shown that there are
race conditions and bugs in the project negotiation so that
a project negotiation returns with exit status OK even if the
remote participant is no longer in the session and so this handler
would fail to handle the leave correctly.

The user is now always asked to leave the session if he is left alone
regardless of the project negotiation status.

Change-Id: I968f69df522af81f803e84ce835a67fd1b527529
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/585
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>