[FIX] Wizard closing dialog
After some testing I found out that this dialog will be opened
if the close() method is called on the container but it
currently executed operation via the run method.

GUI changes:

The information and error dialogs are now shown after the wizard
finished.

This wizard will always close, regardless of the exit status or
thrown exceptions.

It no longer possible to cancel the invitation after it has been
accepted (just the GUI part).

Change-Id: I93ab25d6a374b79e6357c9f5bfc9be6d6b8dc400
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/586
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>