[FIX] fixed non modal dialog handling
The AddProjectToSessionWizard is a non modal wizard but
it was set to block when it should be opened. If a 
modal shell is currently open when this wizard is about
to open it crashed the whole GUI in that way that
is was not possible to close the modal shell.


Change-Id: I2f74f1b7b8f92b7aed20a09fe9a6df50295b267a
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/591
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>