[INTERNAL] removed user invitation completed flag
If you are looking carefully at the session negotiation
you will find out that the user is added, the
user list is synchronized and just after the synchronization
the user completed flag is set and the user list is
synchronized again.

It seems that this is a left over from a very old
negotiation protocol where the users session
invitation is completed after it had received
the whole project.

Change-Id: I55b1cfbfdcabc736672c621f583389bcbb2956fd
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/596
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>