[STF] fixed a test case
Commit a4fca5068efe66696702e0f9d93be6e8f287e6ca introduced 
a small delay when the leave packet should be send to the host.

This test case assumed host left = peer left which is no longer
the case and as the host was still in the session and so not
listed in the roster, the STF tried to add that contact.

Change-Id: I9606f5ea235f68cb09f3dfea20b7af13472cb7a4
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/625
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>