[INTERNAL] release non used threads
The Socks5Transport is mostly used at the beginning of
a session and seldom during a running session.

The current implementation used a fixed thread pool and
so wasting resources because they are not needed as
most threads will always be in a idle state.

Replaced the FixedThreadPool with a CachedThreadPool
that will release non needed threads if they become
idle for more than 60 seconds.


Change-Id: I8e7eab756f7ab9665a2b98a5c802b31763a49b9d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/644
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>