[INTERNAL] simplified packet reassembling
The original packet reassembling was more than complex.

Reason: The whole project archive was not streamed but
send as one packet (which was fragmented during transmission)

As it should be possible to cancel the file transfer during
a project synchronization the possibility to cancel the transfer
was implemented into the network layer and became obsolete
with commit 918f9026b752cf96b211c776ed76a9438f3d7bf7 and
so commit ac9165cd677e268fa9c0c8318b380066d0a5fcee and
92e6de9812128bee351ccd4045dcb9af3bca4ef9 removed all
artifacts that were needed to implement this behavior.

This patch removed the last artifact: the reassembling
of packets fragment in an external thread.

The packet fragments are now reassembled in the
BinaryChannel and the TransferObject is only forwarded 
if all fragments of a packet have arrived.

Change-Id: I08d9ec2f33f2f8119c6e7c5f3d739fdc9e94774f
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/647
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>