[API] removed the isReceiving method from the DataTransferManager
This method was only used by the ActivitySequencer to "guess"
if it should wait a little bit longer until to discard 
activities.

As nobody knows how long the transmission may take this
logic part does not solve anything.

Just increased the timeout as it would be if the
DataTransferManager would currently receive data.

Last but not least the transfer of activities via 
chat is currently disabled so it is currently impossible
to receive activities with non linear sequence numbers.

Change-Id: I2af8728773bec4ea4b7a6a687851b2806f0131c9
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/649
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>