[INTERNAL|GUI] removed GTALK account hack
Due to a bug in the Smack API the username must
be altered so that a successful login is possible.

As it is not desired to create hacks / workarounds for
third party libraries that are actually maintained this
hack has been removed. This should add more pressure towards
the Smack developer to fix their library.

It is still possible to connect via GTALK by just disabling
the SASL authentication.

As for GUI changes ... a best effort is now made to detect why
the login to the server has failed. This currently requires 
parsing the message of the XMPPException but Smack currently
offers no better way as in most cases the returned XMPPError
is always null.

Also inlined various strings again as they were poorly
externalized. These strings should be externalized again
in a further patch.

Change-Id: Iea8712fab0c07af0cf96c73238a18ef3de397a57
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/659
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>