[HACK] #796 Default favorite color is not assigned randomly
This bug is not really a bug but it can result in sub optimal
color distributions if the user do not change its favorite
color after upgrading Saros to version 13.3.* and later.

This hack will assign a random favorite color on the first startup
if the favorite color is invalid / unknown.

Change-Id: Icf235d6520461ee047962ec45de2fde087e3b198
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/665
Tested-by: Jenkins CI
Reviewed-by: Arsenij Solovjev <xeper000@gmail.com>
Reviewed-by: Patrick Schlott <pschlottk@gmail.com>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>