[FIX] disposing a Saros Session resulted in and endless loop
Disposing a Saros Session resulted in and endless loop because
the Session instance itself was put into the container and also
implemented the Disposable interface.

Luckily the PICO Container is smart enough to detect those
errors and throws a already disposed exception.

Merged the stop and dispose methods and removed the isStopped method.

TODO: start and stop methods should not be part of the public interface
and so only be used by the SarosSessionManager.

Change-Id: Ifc2e999122af27e8de83f12aed2292fb68dfc6f9
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/715
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>