[INTERNAL] increased robustness during connection disconnect
The DatatransferManager just uninitialized the transports
regardless whether a current connection establishment
was in progress. This lead to NPE that can be found
in various log files from Saros user.

Made the logic more robust, so it will now throw IOExceptions
instead of NPE which have no meaningful value in the log files.

Change-Id: I47432abfb2fa36e4e1678705b23c4d3300153e6b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/755
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>