[HACK] increased delay until first ping is send
The session on the invitee side is started before
the host have added the invitee to the current session.

Sending activities in this state may lead to stalls
in the ActivitySequencer on the host and other
clients side.

Increased the delay until the first ping is send to 2
minutes which is actually longer than the current
session negotiation timeout (1 minute).


Change-Id: I42c14162f3d2361bfb7b11c8673b5c58ff1f6896
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/765
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>