[FIX] destroying the chat can freeze the GUI
Found this issue while looking at the duration of 
the STF regression. After updating SMACK the
regression time went up from 30 minutes to 
45 minutes.

It seems that SMACK now closes the connection a
little bit slower, therefore the destroying the
chat may wait for a response packet that will
never be received.

This patch will wait until (at least 5 seconds)
the chat is correctly destroyed to ensure that
the connection is not closed in the meantime.

Change-Id: Ib278d08ba5656fce2ae56d48263531126db8ae44
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/779
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>