[HOTFIX] VCS was not working
Commit b7a3c5e650f87d6d1fd64019d8da0113058ef7b6
moved the file list calculation from the
OutgoingSessionNegotiation to OutgoingProjectNegotiation class.

For an unknown reason I replaced sarosSession.useVersionControl()
with the userVersionControl variable that was already
present in this class. Unfortunately this variable
was only initialized once in the CTOR (false) and
then never written again and so broke the VCS
feature.

Change-Id: I76ca1b831329e76894f0af155226262c325ccad4
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/786
Tested-by: Jenkins CI
Reviewed-by: Patrick Schlott <pschlottk@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>