[INTERNAL] removed ConnectionTestManager component
The only useful "feature" it offers to the end user
is to sniff up the current IP address of a remote
contact.

Removed the component for security reasons.

Developers also do not need this component
as the log files already provided enough
information to see which connection mode
is used and last but not least will see
it in the AddProjectsToSession dialog.


Change-Id: I691efd566a51162aa62fb000d64fd33e8ae6037e
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/610
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>