[INTERNAL] removed Ping-Pong Feature
This feature uses activities to generate Ping/Pong
results. So those values are not accurate as the
SWT Thread might been blocked. Furthermore this
feature has no usage. Nobody is looking at
the log files where the Ping/Pong results
are populated.

Conclusion: removed the feature as it should
be replaces with a KeepAliveLogic which
should use PacketExtensions and direct
established connections to determine
if connection is broken and so
abort / kick the user from the current
session.

Change-Id: I60e171cac687f964625348e8a233e9ddd241d447
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/790
Tested-by: Jenkins CI
Reviewed-by: Patrick Schlott <pschlottk@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>