[FIX] user color square is no longer hidden
When selecting a user in the SarosView (running session) the
color square was not visible (Win XP) or shaded (Win Aero).
This patch will draw the color square on top of the current
background to ensure that it is fully "visible".

See also https://sourceforge.net/p/dpp/feature-requests/83/
for the old session view design.

Change-Id: I278d81804dc6e110538c6b082b510b101a29a0d8
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/798
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>