[GUI] removed random follow mode
Removed the random follow mode from the follow mode button
in the toolbar.

The button is now always enabled if the session do not
contain more than 2 users and will always follow the
remote side.

The button will be disabled for sessions with more than
2 users and will only be activitated if a user is selected
in the session overview which than can be followed when
pressing this button.

Change-Id: I089b0378f49a6a469d64743b452ad8a15b30cc66
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/796
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>