[REFACTOR] Split methods in Outgoing/IncomingSessionNegotiation
In preparation for better "observability" of the whole invitation
process (as a precondition for better modularization), this patch splits
some methods on both sides of the SessionNegotiation.
One goal was to separate methods that actively perform a step and those
which only wait for changes on the other side.
Furthermore, some methods were renamed to highlight there actual purpose
(e.g. "awaitRemoteSessionNegotiationData" -->
"awaitClientSessionPreferences", or "modifySessionNegotiationData" -->
"determineSessionParameters").
Finally, instead of "peers" these two classes now speak of "client" and
"host" where applicable.

As a minor addition, the methods create/deleteCollectors were grouped,
and the overriding methods were moved to the top of the classes.

Change-Id: Ifd39c4e8cd7546417c69d3ea87b8fa3e33c4085c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/826
Tested-by: Jenkins CI
Reviewed-by: Patrick Schlott <pschlottk@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>