[REFACTOR] Removed chat and color logic from the SessionNegotiation
Chat and color specific information is still exchanged during the
SessionNegotiation. But from now on, there are
"SessionNegotiationHooks" which may hook into the different stages
of the invitation in order to transfer their specific information. This
way the Incoming/Outgoing processes do not need to know which additional
features Saros supports.

The following dependencies were removed:
 * IncomingSessionNegotiation -> PreferenceUtils
 * InvitationParameterExchangeExtension -> MUC
 * InvitationParameterExchangeExtension -> Colors
 
Through the last two, indirectly the dependencies between
Incoming/OutgoingSessionNegotiation and MUC/Colors were removed. (At
least almost: There is a color related HACK, that should be removed for
clarity.)

This patch did not touch the logic within the extensions (although the
temptation was high). For instance, consider the SarosSessionManager
which needs to know about the user's color (this dependency was removed
from the client side CTOR; otherwise the hooks wouldn't have worked).

However, this patch still breaks the compatibility with older versions
because the additional data during the SessionNegotiation is transferred
in a new format.

Change-Id: Ib14d606958413a5b364831d9f6c6b0921ebeb1a8
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/829
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Tested-by: Jenkins CI