[INTERNAL] SessionManager now updates the NegotationObservables
The de- and registration of negotiation processes where done
in the classes itself. This already led to failures because
the negotiations were added before the CTOR was fully
executed and was already fixed in a previous patch.

This patch passes the handling to the SessionManager which
is already responsible for managing those negotiations to
avoid possible errors in the future.

Change-Id: I2e7b54b0d7601cc479e38412d2c18543efb18b2e
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/830
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI