[INTERNAL] Implemented Client-Server-Architecture for Activities
This patch changes the way we send Activities from mixed
P2P/Client-Server to Client-Server for all Activities.

Clients now send their Activities only to the server on the host-side.
The server is responsible for distributing the Activities to the right
clients.

Also changed the mentions of ActivityDataObjects in CDC and CDS with
Activities, as we don't work with ADOs here.

TODO TextAddedActivities avoid the ConcurentDocumentClient and therefore
break the Client-Server-Architecture.

Change-Id: I6b9a3daac120ea3d8677be3cfe219ee8097a1f05
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/839
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>