[INTERNAL] Read-Only-Users now keep their Jupiter-Documents
Before this patch the Jupiter-Documents of a client that was restricted
to Read-Only-Access was removed on both the client- and the server-side.
This resulted in some special-cases where different Activities were send
to participants with RO-access and therefore had to be handled
differently.

Read-Only-Users now keep their Jupiter-Documents and are treated like
all
other users to unify the Activity-Sending-process.

As we can manage a session with everyone having write-access there is no
real reason to "fasten" this process for RO-User by including many
special cases.

Also changed mentions of ActivitiyDataObjects to Activities in the
CDC(incoming)

Change-Id: Ifc2ebc942510645d6f9d5adcffa2f27bef8281c4
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/840
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>