[INTERNAL] Removed ProjectAddedActivity
Removed the ProjectAddedActivity and associated Classes. This Activity
was the only one used during the Negotiation phase and was not handled
like other Activities as it was shortcut around most Components that
were responsible for handling Activities.

An OutgoingProjectNegotiation now sends a XMPP-Message using the
ProjectNegotiationOfferingExtention and incoming PNs get handled by the
InvitationHandler to match the Session Negotiation Process.

Change-Id: I3c8b298fa5d1bb9bd634e82e1b2a411449b1567c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/851
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>