[INTERNAL] Eliminated AbuseActivityCreated
The sarosSession.AbuseActivityCreated() method was used to create
Activities directly from SarosSession. As Activities should be created
by the responsible ActivityProvider this method was already @deprecated
before.

The only remaining reference to this method was the creation of
permissionActivities. This is now handled by the PermissionManager
that's also responsible for executing incoming permissionActivities.

Also the PermissionManager now dispatches activities to its internal
ActivityReceiver.

Change-Id: Ic0324c94aa7aabf1dbeb3d8030453a321d7d27f0
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/852
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Tested-by: Jenkins CI