[FIX] Eliminate TargetedActivityWrapper
Commit 9635fee (change to Client-Server) broke the execution of
Activities transformed into ITargetedActivities by the
TargetedActivityWrapper. Activities wrapped this way lost the
target-information when transformed into ADOs.

All occurrences of the TargetedActivityWrapper where removed and
Activities using it now either are real ITargetedActivities
or send to everyone as the target was everyone to begin with.

The FileActivity was split into FileActivity and RecoveryFileActivity.
FileActivities have a special purpose during the Recovery-Phase and this
is the only point where FileActivities need to be send to specific
users.

Change-Id: Ib761f8e1049b8bac3ad6c563e4ceae2a0230291a
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/859
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Tested-by: Jenkins CI