[INTERNAL] ITargetedActivities now only has one Target
Changed the ITargetedActivity-Interface. The getRecipients() that
returned a List<User> was changed to getTarget() that returns a single
User. This change was due to the fact, that we have no Activity with
more than one target and the getRecipients()-method always returned a
single User transformed into a List<User>.

Change-Id: I8df0d7716d4d78b183df0664b7bd6b9d12b8790d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/861
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>