[INTERNAL] Simplified numbering mechanism for outgoing activities
The numbering of outgoing activities (wrapping them in
(TimedActivityDataObjects) stretched across three classes. Futhermore,
this implementation blurred the border between the handling of incoming
and outgoing activities. Outgoing is pretty simple and has no need for
three classes: It is now done by the ActivitySequencer alone. The
ActivityQueueManager (and the ActivityQueues) are now solely responsible
for incoming activities.

Change-Id: Id2866ab0c22e23f3ca25b69e95c36d670c38ad90
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/867
Tested-by: Jenkins CI
Reviewed-by: Patrick Schlott <pschlottk@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>