[UI] Disabled Non-Host adding projects to session
Disabled the possibility for non-hosts to add projects to the session.
Projects that were added by non-hosts could cause strange behaviors,
especially if the host delayed or canceled the Negotiation.

Additionally we want to incorporate a client-server-architecture and
this patch further advances in this direction.

This feature should be "reactivated" using the
client-server-architecture. Meaning, that the client sends the project
to the host and he sends them to everyone else.

This patch only deactivates the UI-Part of this feature, so that it can
not be accessed by non-hosts.

Change-Id: Id487a83c58f13534c14c92aa41a5697918221696
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/870
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Bernd Bieber <brotlee@zedat.fu-berlin.de>