[STF] Updated test case and added RegexMatcher for TreeItems
Updated the test AwarenessInformationsVisibleAfterInvitationTest by
synchronizing on the ActivityQueue instead of waiting a fixed time.
Added a second test case which tests if a participant also sees
awarenessInformation of other participants if he joins a running
session.

Also added and used the getNodeWithRegex()-method for TreeItems to find
the Participants in the SessionComposite.

Change-Id: Iee9c4789d9dd3bd85faace60acec604af3069b55
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/877
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Patrick Schlott <pschlottk@gmail.com>