[FIX] Transmit switches to editors of non-shared-files
Since commit 609885ac8a6518b3c9c3771560989cf2d5ff2dcb only
IResourceActivities that relate to a file that the
remote user is ready to receive will be transmitted to them. This
resulted in EditorActivities about non-shared-files (such as switching
to the editor of such a file) to produce an NPE and not be sent as their
path is NULL.

This resulted in remote users not seeing that the local user opened a
non-shared-file.

IResourceActivities that have NULL as path will now be treated by the
ActivityHandler as being not resource-related as we can't decide whether
to sent them or not.

Change-Id: Ic50e1f12b45bfb45e2fb684ad81edcfd2db2a60b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/885
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI