[INTERNAL] removed activity dropping in the activity sequencer
I do not know why this was implemented. Activities were
dropped when they not arrived in a given time period.

This can currently only happens if the transmission of
activities fails. There is currently just a log statement
that informs about this error. Furthermore it is possible
that a connection will be established again afterwards which
may result in a completely broken session.

This patch ensures that the session will at least hang forever
because it is out of sync.

This patch also fixed an issue with sequence number overflows
which was not handled correctly and so made it impossible to
use the Saros Plugin as "standalone" server.

Change-Id: I786b7ebf03d14b4501e224ee4a0be7a82571ce11
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/895
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>