[INTERNAL] removed unused methods from User class
With commit c80a498be7b5c3032bf65c824eb57a57d01ae444 all
methods the returns the user network status became obsolete.

As the commit message already states the network layer should not
know anything about "Users" and a "Session" and so the detection
of broken network connection should not be done with such "helper"
methods in a class that has nothing to do with the network layer.

The methods in the XMPPTransmitter already throw IOExceptions which
should be enough to detect that a stream / connection to another user
is broken.

Change-Id: I20982b09e68c0d785156f8a0441b0d5b875b8e05
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/909
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>