[INTERNAL] cleaned up SarosSession class
Merged code from CTORS and initializeSessionContainer method so that
it is possible to declare much of the used variables as final.

Removed senseless @inheritDoc with no further comments.

Removed NLS format from an exception message.

Replaced invocation of No-Argument CTOR of exceptions with
Message CTOR.

Moved the internal Javadoc for getResourceQualifiedJID method
to the ISarosSession interface.

Change-Id: I7930108167aba5fde213091fef22f973fcdfb703
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/920
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>