[FIX] ActivityUtils.optimize swallowed DAOs and changed order
This utility method tries to optimize the amount of DAOs that
should be send.

Bug 1. Only one selection DAO was returned also it is possible
to have multiple selections (one for each editor).

Bug 2. The viewport and selection DAO were appended to the
end of the list which may result in wrong awareness status
at the remote side.

Bug 3. The delete folder DAO optimization may produce wrong results
depending on other DAOs that are interleaved between them. Removed the
optimization of delete folder DAOs as the SharedResourceManager already
sends the minimal amount of possible FileActivities on resource changes.


Change-Id: I3d6bca2a1c0a805b70434b77147e70801dc42fb7
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/977
Tested-by: Jenkins CI
Reviewed-by: Patrick Schlott <pschlottk@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>