[FIX] #798 PreferenceManager does not modify shared projects
Removed this component as it is HIGHLY dangerous because it
manipulate/changes encoding and line delimiters for
a shared project (which is luckily not working) WITHOUT asking
the user. Even if those changes are reverted at session end does
not mean:

1. that files that were affected by those changes are correctly "rolled
back"

2. if the session or Eclipse crash you are really out of luck and
furthermore you would not even know the Saros made changes to the
project setting(s).

TODO: exchange the relevant information in the OPN/IPN and let the USER
decide how to deal with those incompatibilities.

Change-Id: I53100a34649e69602a7c90f91cbba64e2c8f0f08
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/978
Tested-by: Jenkins CI
Reviewed-by: Patrick Schlott <pschlottk@gmail.com>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>