[FIX] EnterXMPPAccountComposite returned wrong dimensions for layout
An ExpandableComposite will only tell its full height if it is expanded.
Since this composite is not expanded by default we must ensure that the
correct size is set. This is simply achieved by asking the composite to
calculate its default size and set this size in the layout data.


Change-Id: I50fe7f6d8dcf8c581a7372cdef7b6fba123c9c02
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/983
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>