[JUNIT] corrected InviteWithDifferentVersionsTest test case
The property for displaying the the needed error shell was
renamed and so this shell was not shown anymore.

Corrected the test case in such a way that is now testing
the version mismatch in normal user mode instead of debug mode.

It now assumes that the session negotiation fails instead of showing
an dialog that let the user allow to bypass the version negotiation.

Change-Id: I85b8c713a74707de320ae800d9ea4a19a6f2306c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1031
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>