[INTERNAL] removed dialogs for incompatible version handling
It is no longer possible to bypass the version check since
Saros version 13.3.22 and so those dialogs were only displayed
if the property was manually set. As this should be only done
by developers they will mostly now the results of this change
and so do not need additional dialogs that telling them that
there is a version mismatch.

Change-Id: If7cf69d46e72684e9fd37b64d9d486659d960a88
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1038
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>