[INTERNAL] improved user list synchronization
It is possible that multiple users leave at the same time. This may lead
to synchronizations where nothing is to synchronize and so the
synchronization just timed out. While this behavior has no real impact
on the session (users are just removed with a little delay) the STF uses
a hard timeout which was exceeded in those situations.

User list synchronization will now abort prematurely if a user that is
part of the synchronization left the session in the meantime.

Change-Id: Ib3d84e5ec0d6fd34e366d7f404476c039a314d8f
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1043
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>