[INTERNAL] fixed NPE and changed cancel notification for IPN/OPN
The getUser method already states that it MIGHT return null.

It is possible the a user may leave the session during IPN/OPN.

The cancel notification is now send via stream and only falls back
to XMPP chat if that fails. This should ensure that the cancel
notification is received before the leave session notification arrives.


Change-Id: I99e1cf4b80e05634d4f3685aba3d09f6a7de77ea
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1071
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>