[FIX] Jupiter documents were registered to early for clients
This is a partial fix for the following problem:

The remote users were registered to early on the host side at the
JupiterServer so following happened:

Alice invites Bob and Carl. While Carl is still in project negotiation
Bob has already finished the process and starts working. The
JupiterServer will now register Carl for documents that Bob is currently
modifying but Carl will currently not receive those Jupiter activities.
The result is when Carl finishes the project negotiation he will receive
Jupiter activities with timestamps that are *far* in the future eg.

Alice sends (X, 0) but Carls Jupiter client is at (0, 0) and so refuses
to apply the transformation.

TODO: this does not work for any projects that are post shared in a
running session.

Change-Id: I7687e7af699b8dca439ed5634a45f5ff5bfeea53
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/1126
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>